components/pmsx003: Add support for missing values from sensor
The PMSx003 sensors not only return AQI indicators for PM1, PM2.5,
and PM10 but also the standar PMs (in addition to the currently reported
values which are the values under atmospheric pressure, as well as the
particle counts by size.  The later values being in decilitres.  This adds
in the rest of those data points from the sensor for tracking by
esphome.

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>